requirements.txt
asgiref==3.3.1
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.1
dj-rest-auth==2.1.3
Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
idna==2.10
oauthlib==3.1.0
Pillow==8.1.2
pycodestyle==2.7.0
pycparser==2.20
PyJWT==1.7.1
python-dotenv==0.15.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
stripe==2.56.0
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
